+++ packages/aggregation/src/index.ts
@@ -1,3 +1,4 @@
 export * from './signature'
 export * from './dedupe'
 export * from './priority'
+export * from './upstream'