+++ packages/app/src/index.ts
@@ -1,5 +1,6 @@
 export * from './runtime'
 export * from './oplog'
+export * from './paging'
 export * from './repairindex'
 export * from './http'
 export { buildServerDeps, startServer } from './server'